Fix missing help pages resulting from non-well-formed markup.
Most of the special case asciidoc replacements were also removed since
they're a bit confusing and doing so results in a 60% improvement in the
time taken by asciidoc to process the help pages.  Unfortunately, I'm
now limited to preparing a three course menu for my guests while it runs
rather than the degustation extravaganzas of the past but such is life.

It's still a bit of a mess but since we're almost certainly moving to
something else in the near future it's probably not worth cleaning it up
before 2.0 is released.