Added "filetype=asciidoc" to help file modlines.
At the moment, only "syntax=asciidoc" was specified. The
asciidoc_filetype.vim that is distributed with Asciidoc gives the
installer two options -- to either associate every *.txt file with
asciidoc or to try to decipher whether or not the file is an asciidoc
using the first 50 lines.

So it's not given that a developer will open the *.txt files with
ft=asciidoc, and so it's helpful to have it explicitly in the modeline.