In delmarks, change let variable name to prevent namespace-related error (bug #126).
See http://vimperator.org/trac/ticket/126

Using...

    function (args)
    {
        let args = args.string;
    ...

was causing Firefox to complain that args was undefined. Changing the
"let" to "var" or changing the names of the variables (e.g., changing
"function (args)" to "function (arg)") fixes the problem. IMO, this
appears to be a bug in Firefox.

The short fix would be to change the formal argument from "args" to
"arg." However, that would make delmarks' implementation different from
all the rest. So changed "let args" to "let argstring" and adjusted
throughout the function.