Fix S- handling to normalize in the manner of vim
Instead of treating all key presses as Shifted, it now normalises the
user-input.

Also now treats < as <lt>

Behaviour:

(:map <C-S-a> == :map <C-S-A>) != (:map <C-a> == :map <C-A>)
:map <C-S-@> is meaningless (though no error is produced) c.f. vim
:map <C-S-2> is meaningless (though no error is produced) c.f. vim

Internally, commands are stored with no shift modifiers for alphabetic
characters, i.e. (:map <C-S-a> creates the mapping <C-A>) as in vim.

<CapsLock> <C-a> will cause the <C-A> mapping (mapped with <C-S-a> to be
triggered).

There still seems to be a bug with :map <Space> and :map <CR>